move struct V6TrainingData to trainingdata.h
